services/baleal/requirements.txt
python-dotenv
python-telegram-bot
instagrapi
Pillow
moviepy